# entity-service/openapi.yaml
openapi: 3.0.1
info:
  title: Entity Service API
  version: 1.0.0
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /cases/{id}/tasks/search:
    post:
      summary: List tasks for a case (ServiceNow data source only).
      operationId: searchCaseTasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchCaseTasksRequest'
      responses:
        "200":
          description: Tasks for the case.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchCaseTasksResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Case not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cases/{id}/tasks:
    post:
      summary: Create a task on a case (ServiceNow data source only).
      operationId: createCaseTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseTaskRequest'
      responses:
        "201":
          description: Task created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Case not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "503":
          description: Task writes are not yet available (downstream integration not deployed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{id}:
    get:
      summary: Get a task by ID (ServiceNow data source only).
      operationId: getTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The task record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Task not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update a task (ServiceNow data source only).
      description: >
        Updates exactly one of state, assignedToEmail, or dueDate.
      operationId: updateTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        "200":
          description: Task updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Task not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "503":
          description: Task writes are not yet available (downstream integration not deployed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CaseNumberRef:
      type: object
      nullable: true
      properties:
        id:
          type: string
        number:
          type: string
        type:
          type: string
          enum: [case, incident, change_request, problem]
          nullable: true
          description: >
            What kind of record this reference points at. Null when the
            backing data source could not resolve the referenced record's
            kind; a client must not then assume it is a case, since linking
            to a case detail page for a non-case record would 404.
...
    EntityRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
...
    TaskSummary:
      type: object
      required: [id, subject, updatedOn]
      properties:
        id:
          type: string
          format: uuid
        subject:
          type: string
        state:
          type: string
          nullable: true
          description: >
            Plain string, not a closed enum. Known values: OPEN, CLOSED, OTHER
            (fallback for any unmapped raw ServiceNow value), or null.
        dueDate:
          type: string
          nullable: true
        assignedTo:
          $ref: '#/components/schemas/EntityRef'
          nullable: true
        updatedOn:
          type: string
    SearchCaseTasksRequest:
      type: object
      properties:
        pagination:
          type: object
          properties:
            offset:
              type: integer
              minimum: 0
              default: 0
            limit:
              type: integer
              minimum: 1
              maximum: 50
              default: 20
    SearchCaseTasksResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskSummary'
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
    CreateCaseTaskRequest:
      type: object
      required: [subject]
      additionalProperties: false
      properties:
        subject:
          type: string
        dueDate:
          type: string
          format: date-time
        assignedToEmail:
          type: string
          format: email
        visibleToCustomer:
          type: boolean
    UpdateTaskRequest:
      type: object
      additionalProperties: false
      description: >
        Exactly one of state, assignedToEmail, or dueDate must be provided.
      oneOf:
        - required: [state]
        - required: [assignedToEmail]
        - required: [dueDate]
      properties:
        state:
          type: string
          description: >
            Plain string, not a closed enum — ServiceNow may surface undocumented
            raw values; do not assume a closed set.
        assignedToEmail:
          type: string
          format: email
        dueDate:
          type: string
          format: date-time
    TaskDetail:
      type: object
      required: [id, subject, visibleToCustomer, createdOn, updatedOn]
      properties:
        id:
          type: string
          format: uuid
        subject:
          type: string
        state:
          type: string
          nullable: true
          description: >
            Plain string, not a closed enum. Known values: OPEN, CLOSED, OTHER
            (fallback for any unmapped raw ServiceNow value), or null.
        dueDate:
          type: string
          nullable: true
        visibleToCustomer:
          type: boolean
        assignedTo:
          $ref: '#/components/schemas/EntityRef'
          nullable: true
        requestType:
          type: string
          nullable: true
          description: >
            Raw ServiceNow choice-list key. Plain string, not a closed enum,
            since the set of valid keys is SN-configuration-dependent.
        requestTypeLabel:
          type: string
          nullable: true
        environment:
          type: string
          nullable: true
        environmentLabel:
          type: string
          nullable: true
        product:
          $ref: '#/components/schemas/EntityRef'
          nullable: true
        parentCase:
          $ref: '#/components/schemas/CaseNumberRef'
          nullable: true
        createdOn:
          type: string
        updatedOn:
          type: string
...